Refactor: winResult 관련 로직 class 분리
