+++ src/main/java/lotto/Application.java
@@ -5,6 +5,6 @@ import lotto.service.LottoGame;
 public class Application {
     public static void main(String[] args) {
         LottoGame lottoGame = new LottoGame();
-        lottoGame.game();
+        lottoGame.playGame();
     }
 }
